реструктуризировано с учетом изменений
