+++ frontend/app/(auth)/layout.tsx
@@ -1,0 +1,17 @@
+import { ReactNode } from 'react';
+
+interface AuthLayoutProps {
+    children: ReactNode;
+}
+
+export default function AuthLayout({ children }: AuthLayoutProps) {
+    return (
+        <div className="min-h-screen bg-gradient-to-br from-blue-50 to-indigo-100 flex items-center justify-center p-4">
+            <div className="w-full max-w-md bg-white rounded-xl shadow-lg overflow-hidden">
+                <div className="p-8">
+                    {children}
+                </div>
+            </div>
+        </div>
+    );
+}
